compare selected folders
