+++ package.json
@@ -106,6 +106,10 @@
         {
           "command": "foldersCompare.compareFoldersAgainstWorkspace",
           "when": "workspaceFolderCount != 0"
+        },
+        {
+          "command": "foldersCompare.compareSelectedFolders",
+          "when": "false"
         }
       ],
       "view/title": [
@@ -113,6 +117,12 @@
           "command": "foldersCompare.refresh",
           "group": "navigation",
           "when": "view == foldersCompareAppService && foldersCompareContext.hasFolders"
+        }
+      ],
+      "explorer/context": [
+        {
+          "command": "foldersCompare.compareSelectedFolders",
+          "when": "explorerResourceIsFolder && listDoubleSelection"
         }
       ]
     },
@@ -137,6 +147,10 @@
           "light": "resources/light/refresh.svg",
           "dark": "resources/dark/refresh.svg"
         }
+      },
+      {
+        "command": "foldersCompare.compareSelectedFolders",
+        "title": "[Compare Folders] Compare selected folders"
       }
     ],
     "viewsContainers": {
